moved all camera viewer controls out into separate component to clean things up a bit
