+++ frontend/public/index.html
@@ -37,7 +37,7 @@
       <script>
       window.sensei = {
         env: {
-          baseUrl: 'http://localhost:5007',
+          baseUrl: 'http://localhost:5001',
           tcBaseUrl: 'http://localhost:3000/',
           production: false
         }
